fix: nit

+++ nexus/catalog/migrations/V40__maintenance_flows.sql
@@ -2,8 +2,8 @@ CREATE SCHEMA IF NOT EXISTS maintenance;
 
 CREATE TABLE IF NOT EXISTS maintenance.maintenance_flows
 (
-    id              bigint PRIMARY KEY,
-    flow_id         serial    NOT NULL,
+    id              serial PRIMARY KEY,
+    flow_id         bigint    NOT NULL,
     flow_name       TEXT      NOT NULL,
     workflow_id     TEXT      NOT NULL,
     flow_created_at TIMESTAMP NOT NULL,
